The @ key is now automatically added when creating a comment, the input field for this now also uses the Twitter Bootstrap's input group to display the @ symbol before the text field.
Updated the add hours SQL query to include an on duplicate key, now adds hours together for the same day instead of multiple rows for the same day.

My version of the database is slightly different now as well, with the date column in the work table now being a unique key.