Starting creating a search class.
This all kind of happened because I was all like "I feel like making a search class, screw planning and logic, lets see what I can create at 10:30pm at night".

Well after spending a good 3 and a half hours on my bum, I present a semi-working easy to use use class.

How you use it is you can go:

$testingSearch = new TaskSearchHelper();
    	$testingSearch->setTag("addedHours", false);
    	$testingSearch->setTask(127, true);
    	$testingSearch->setUser(2, true);
    	$testingSearch->setDatabase($this->database);
    	$testingSearch->search();

Because tags/comments are the one ones with both a task id and a user id attached to the table, the class will only search for comments, if you only entered in a task it will search for tasks, etc.

You are also able to set the page number and quantity (with the page number being where to LIMIT the SQL query).

I do warn you though, I didn't really think through the variable names and I decided to experiement and see if I could come up with something awesome. However, I did not.

With that being said, I tried to keep it all nice and open and flexible, but then again it's 2:13am so I probably have no idea what this code actually does or looks like and I'll wake up and it will all be a big pile of mush.

Anyway, I'll continue this as my own little probject to try and get us a few bonus marks on the task system for our team website.

Enjoy!