Fixed a few minor bugs.
Now when you create a new Task, the user's name is displayed without loading the page (originally nothing was there until you reloaded the page).

Now when you create a new Task, the Task description will display properly with the option to collapse the text (before the "..." was showing up until you reloaded the page).

Added a nice and dirty way around formatting the TaskSearchHelper, I have left the following comment in the method buildFromQueryRow inside Task:

/*
 * Please note: I am setting the Task's title to be an array instead of just the task title
 * so it formats nicely when printed out in the search field.
 *
 * It might be worth making a different method to return a nicely formatted array
 * for the TaskSearchHelper.
 */

Creating that seperate method is something I intend to do, but is fairly low on the priority list.

Also, the date picker is now the JQuery date picker instead of the type="date" html date picker as the original way only worked with Chrome and I think Safari.